requirements.txt
streamlit==1.33.0
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.34
langchain-huggingface==0.0.3
pypdf==4.1.0
python-docx==1.1.0
docx2txt==0.8
faiss-cpu==1.8.0
sentence-transformers==2.2.2
python-dotenv==1.0.1